+++ src/plugins/rssyl/parse822.c
@@ -135,7 +135,7 @@ FeedItem *rssyl_parse_folder_item_file(gchar *path)
 				/* Last-Seen timestamp */
 				if( !strcmp(line[0], "X-RSSyl-Last-Seen") ) {
 					ctx->last_seen = atol(line[1]);
-					debug_print("RSSyl: got last_seen timestamp %ld\n", ctx->last_seen);
+					debug_print("RSSyl: got last_seen timestamp %lld\n", (long long)ctx->last_seen);
 				}
 
 				/* ID */
